requirements/common.txt
Django==2.0.6
psycopg2-binary==2.7.5
graphene==2.1.2
graphene-django==2.0.0
django-graphql-jwt==0.1.10
